feat: added new authentication flow

+++ internal/domain/category.go
@@ -13,12 +13,12 @@ type Category struct {
 
 // CategoryRepository represents the categories repository contract
 type CategoryRepository interface {
-	GetByID(ctx context.Context, id int64) (Category, error)
-	GetByUserSUB(ctx context.Context, sub int64) ([]Category, error)
+	GetByID(ctx context.Context, id uint) (Category, error)
+	GetByUserSUB(ctx context.Context, sub uint) ([]Category, error)
 	// GetAll(ctx context.Context) ([]Category, error)
 
 	// CreateOrUpdate(ctx context.Context, cat *Category) error
 	Update(ctx context.Context, cat *Category) (*Category, error)
 	Create(ctx context.Context, cat *Category) (*Category, error)
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id uint) error
 }
